+++ examples/cis2-wccd/src/lib.rs
@@ -33,6 +33,10 @@ const TOKEN_ID_WCCD: ContractTokenId = TokenIdUnit();
 /// The metadata url for the wCCD token.
 const TOKEN_METADATA_URL: &str = "https://some.example/token/wccd";
 
+/// List of supported standards by this contract address.
+const SUPPORTS_STANDARDS: [StandardIdentifier<'static>; 2] =
+    [CIS0_STANDARD_IDENTIFIER, CIS2_STANDARD_IDENTIFIER];
+
 // Types
 
 /// Contract token ID type.
@@ -62,7 +66,10 @@ struct AddressState<S> {
 #[concordium(state_parameter = "S")]
 struct State<S> {
     /// The state the one token.
-    token: StateMap<Address, AddressState<S>, S>,
+    token:        StateMap<Address, AddressState<S>, S>,
+    /// Map with contract addresses providing implementations of additional
+    /// standards.
+    implementors: StateMap<StandardIdentifierOwned, Vec<ContractAddress>, S>,
 }
 
 /// The parameter type for the contract function `unwrap`.
@@ -90,6 +97,17 @@ struct WrapParams {
     to:   Receiver,
     /// Some additional bytes to include in a transfer.
     data: AdditionalData,
+}
+
+/// The parameter type for the contract function `setImplementors`.
+/// Takes a standard identifier and list of contract addresses providing
+/// implementations of this standard.
+#[derive(Debug, Serialize, SchemaType)]
+struct SetImplementorsParams {
+    /// The identifier for the standard.
+    id:           StandardIdentifierOwned,
+    /// The addresses of the implementors of the standard.
+    implementors: Vec<ContractAddress>,
 }
 
 /// The different errors the contract can produce.
@@ -141,7 +159,8 @@ impl<S: HasStateApi> State<S> {
     /// Creates a new state with no one owning any tokens by default.
     fn new(state_builder: &mut StateBuilder<S>) -> Self {
         State {
-            token: state_builder.new_map(),
+            token:        state_builder.new_map(),
+            implementors: state_builder.new_map(),
         }
     }
 
@@ -255,6 +274,24 @@ impl<S: HasStateApi> State<S> {
         from_state.balance -= amount;
 
         Ok(())
+    }
+
+    /// Check if state contains any implementors for a given standard.
+    fn have_implementors(&self, std_id: &StandardIdentifierOwned) -> SupportResult {
+        if let Some(addresses) = self.implementors.get(std_id) {
+            SupportResult::SupportBy(addresses.to_vec())
+        } else {
+            SupportResult::NoSupport
+        }
+    }
+
+    /// Set implementors for a given standard.
+    fn set_implementors(
+        &mut self,
+        std_id: StandardIdentifierOwned,
+        implementors: Vec<ContractAddress>,
+    ) {
+        self.implementors.insert(std_id, implementors);
     }
 }
 
@@ -622,6 +659,62 @@ fn contract_token_metadata<S: HasStateApi>(
     }
     let result = TokenMetadataQueryResponse::from(response);
     Ok(result)
+}
+
+/// Get the supported standards or addresses for a implementation given list of
+/// standard identifiers.
+///
+/// It rejects if:
+/// - It fails to parse the parameter.
+#[receive(
+    contract = "CIS2-wCCD",
+    name = "supports",
+    parameter = "SupportsQueryParams",
+    return_value = "SupportsQueryResponse"
+)]
+fn contract_supports<S: HasStateApi>(
+    ctx: &impl HasReceiveContext,
+    host: &impl HasHost<State<S>, StateApiType = S>,
+) -> ContractResult<SupportsQueryResponse> {
+    // Parse the parameter.
+    let params: SupportsQueryParams = ctx.parameter_cursor().get()?;
+
+    // Build the response.
+    let mut response = Vec::with_capacity(params.queries.len());
+    for std_id in params.queries {
+        if SUPPORTS_STANDARDS.contains(&std_id.as_standard_identifier()) {
+            response.push(SupportResult::Support);
+        } else {
+            response.push(host.state().have_implementors(&std_id));
+        }
+    }
+    let result = SupportsQueryResponse::from(response);
+    Ok(result)
+}
+
+/// Set the addresses for an implementation given a standard identifier and a
+/// list of contract addresses.
+///
+/// It rejects if:
+/// - Sender is not the owner of the contract instance.
+/// - It fails to parse the parameter.
+#[receive(
+    contract = "CIS2-wCCD",
+    name = "setImplementors",
+    parameter = "SetImplementorsParams",
+    mutable
+)]
+fn contract_set_implementor<S: HasStateApi>(
+    ctx: &impl HasReceiveContext,
+    host: &mut impl HasHost<State<S>, StateApiType = S>,
+) -> ContractResult<()> {
+    // Authorize the sender.
+    ensure!(ctx.sender().matches_account(&ctx.owner()), ContractError::Unauthorized);
+    // Parse the parameter.
+    let params: SetImplementorsParams = ctx.parameter_cursor().get()?;
+    // Update the implementors in the state
+    host.state_mut().set_implementors(params.id, params.implementors);
+    Ok(())
 }
 
 // Tests